Split GoldenTiles into separate files to make it easier for AS RAM
